+++ src/settings.ts
@@ -55,7 +55,7 @@ export class Settings {
   dimmingEnd = '7:00';
   dimmingStart = '23:00';
   drawConstellations = true;
-  floatHands = true;
+  floatHands = 'T';
   hidePlanets = false;
   hideSeconds = false;
   hourlyForecast = HourlyForecast.VERTICAL;
@@ -90,7 +90,7 @@ export class Settings {
     this.dimmingEnd = Cookies.get('dimming_end') || defaultSettings.dimmingEnd;
     this.dimmingStart = Cookies.get('dimming_start') || defaultSettings.dimmingStart;
     this.drawConstellations = toBoolean(Cookies.get('draw_constellations'), defaultSettings.drawConstellations);
-    this.floatHands = toBoolean(Cookies.get('float_hands'), defaultSettings.floatHands);
+    this.floatHands = Cookies.get('float_hands') || defaultSettings.floatHands;
     this.hidePlanets = toBoolean(Cookies.get('hidep'), false);
     this.hideSeconds = toBoolean(Cookies.get('hides'), false);
     this.hourlyForecast = (Cookies.get('hourly_forecast') as HourlyForecast) || defaultSettings.hourlyForecast;
@@ -108,6 +108,11 @@ export class Settings {
     this.timeFormat = toTimeFormat(Cookies.get('ampm'), defaultSettings.timeFormat);
     this.updateToHide = Cookies.get('update-to-hide') || '';
     this.userId = Cookies.get('id') || '';
+
+    if (this.floatHands === 'true')
+      this.floatHands = 'T';
+    else if (this.floatHands === 'false')
+      this.floatHands = 'N';
 
     const body = $('body');
 
@@ -129,7 +134,7 @@ export class Settings {
     Cookies.set('dimming_end', this.dimmingEnd, expiration);
     Cookies.set('dimming_start', this.dimmingStart, expiration);
     Cookies.set('draw_constellations', this.drawConstellations.toString(), expiration);
-    Cookies.set('float_hands', this.floatHands.toString(), expiration);
+    Cookies.set('float_hands', this.floatHands, expiration);
     Cookies.set('hidep', this.hidePlanets.toString(), expiration);
     Cookies.set('hides', this.hideSeconds.toString(), expiration);
     Cookies.set('hourly_forecast', this.hourlyForecast, expiration);